backend/pyproject.toml
[tool.poetry]
name = "smart-style"
version = "0.1.0"
description = ""
authors = ["chipanatica <josepaulct@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
torch = "^2.5.1"
pillow = "^11.0.0"
openai = "^1.57.2"
supabase = "^2.10.0"
uvicorn = "^0.32.1"
python-multipart = "^0.0.19"
transformers = "^4.47.0"
accelerate = "^1.2.0"
torchvision = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
